% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heartFailure.R
\docType{data}
\name{heartFailure}
\alias{heartFailure}
\title{Heart failure datasets}
\format{An object of class \code{"list"};}
\usage{
data(heartFailure)
}
\description{
Various data (demo, cell-types and omics datasets) from 58 heart failure patients.
demographics:
hosp3months:
cell-type frequencies:
mRNA:
Holter monitor variables:
protein expression:
}
\examples{
data(heartFailure)
}
\references{
Singh et al., Can J Cardiol. 2019 Apr;35(4):471-479.
(\href{https://www.ncbi.nlm.nih.gov/pubmed/30935638}{PubMed})
}
\keyword{datasets}
