% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omicscentral.R
\docType{package}
\name{omicsCentral-package}
\alias{omicsCentral}
\alias{omicsCentral-package}
\title{omicsCentral: Wrapper Functions for Common Bioinformatics Analyses}
\description{
Set of packages that perform commonly used bioinformatics analyses.
   These include exploratory data anlyses, differential expression analysis, network analysis, 
   data integration approaches, literature mining, and manuscript writing.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://omicscentral.com}
}

}
\keyword{internal}
