% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logo.R
\name{omicscentral_logo}
\alias{omicscentral_logo}
\title{The omics central logo, using ASCII or Unicode characters}
\usage{
omicscentral_logo(unicode = l10n_info()$`UTF-8`)
}
\arguments{
\item{unicode}{Whether to use Unicode symbols. Default is \code{TRUE}
on UTF-8 platforms.}
}
\description{
Use \code{\link[crayon:strip_style]{crayon::strip_style()}} to get rid of the colors.
}
\examples{
omicscentral_logo()
}
