% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{omicscentral_packages}
\alias{omicscentral_packages}
\title{List all packages in Omics Central}
\usage{
omicscentral_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include omicsCentral in the list?}
}
\description{
List all packages in Omics Central
}
\examples{
omicscentral_packages()
}
