% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{omicscentral_update}
\alias{omicscentral_update}
\title{Update omics central packages}
\usage{
omicscentral_update(recursive = FALSE, repos = getOption("repos"))
}
\arguments{
\item{recursive}{If \code{TRUE}, will also list all dependencies of
Omics Central packages.}

\item{repos}{The repositories to use to check for updates.
Defaults to \code{getOptions("repos")}.}
}
\description{
This will check to see if all omics central packages (and optionally, their
dependencies) are up-to-date, and will install after an interactive
confirmation.
}
\examples{
\dontrun{
omicscentral_update()
}
}
