% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda.R
\name{eda}
\alias{eda}
\title{eda}
\usage{
eda(eset, response = NULL, demo = NULL, ncomp = 2)
}
\arguments{
\item{eset}{list of nxp matrices (expression sets)}

\item{response}{variable of any type (optional)}

\item{demo}{- demographics data}

\item{ncomp}{number of principal components to retain}
}
\description{
build various biomarker panels
}
